update flow and temp from i2c

+++ TouchGFX/gui/include/gui/model/ModelListener.hpp
@@ -2,6 +2,9 @@
 #define MODELLISTENER_HPP
 
 #include <gui/model/Model.hpp>
+#include <stdio.h>
+#include <stdlib.h>
+
 
 class ModelListener
 {
@@ -28,7 +31,7 @@ public:
 	virtual void addDatapointFlow(int newDP) {}
 
 	virtual void setTestVal1(float newTestVal1) {}
-	virtual void setTestVal2(float newTestVal2) {}
+	virtual void setTestVal2(uint16_t newTestVal2) {}
 	virtual void setDutyCycle(int newDutyCycle) {}
