+++ TouchGFX/gui/include/gui/screen2_screen/Screen2View.hpp
@@ -15,7 +15,7 @@ public:
     
 
     void updateTestVal1(float newVal1);
-    void updateTestVal2(float newVal2);
+    void updateTestVal2(uint16_t newVal2);
     void updateDutyCycle(int newDuty);
 
     void incrementDuty();
